arch: Code cleanup after refactoring aarch64
