+++ arch/src/aarch64/gicv3.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod kvm {
-    use super::super::gic::kvm::KvmGICDevice;
-    use super::super::gic::{Error, GICDevice};
+    use crate::aarch64::gic::kvm::KvmGICDevice;
+    use crate::aarch64::gic::{Error, GICDevice};
     use crate::layout;
     use hypervisor::kvm::kvm_bindings;
     use std::sync::Arc;
@@ -11,7 +11,7 @@ pub mod kvm {
     type Result<T> = result::Result<T, Error>;
 
     pub struct KvmGICv3 {
-        /// The file descriptor for the KVM device
+        /// The hypervisor agnostic device
         device: Arc<dyn hypervisor::Device>,
 
         /// GIC device properties, to be used for setting up the fdt entry
